Add files via upload
Employee website- Abdullah almalki & Abdullah Alshahrani
Almalki:
I worked on the desktop version of the employee store management system, I made the html and js files for it and the initial css as well.

Alshahrani:
enhanced version of the overall experience, major changes to the css to support mobile version. Error checking and fixing.